Add files via upload
Tiling based on Pisot numbers

Some years ago I looked through the web for tilings based on Pisot numbers. There I found a work of Jean-René Geoffroy [1] (french language) containing a Maple V program which is still available. 
This notebook contains my version of the Maple routines converted into Mathematica code. It is restricted to polynomials of order three which have roots that are Pisot numbers. In order to limit the size all output of the notebook is deleted. So the first action is to evaluate the complete notebook. In my opinion the most interesting feature of this implementation is the ability to create sub-tilings of every tiling. You find some more theoretical background in [2] and [3]. As far as I am aware no Mathematica code for tilings based on beta expansion of Pisot numbers is available in the web. The notebook "Perron Number Tiling Systems" from Roger Bagula [4] can produce the border lines of the tilings below. That is because any Pisot number is a Perron number [5].

Please feel free to mail your comments or interesting add-on. My experience in the English language is limited, so please excuse errors.

Author:	Dieter K. Steemann, Bremen, Germany ©
Created:	2008, Jan. 29 with Mathematica 5
Changes:	2012, Nov., 28; converted into Mathematica 7, translated German explanations into English language.
		2014, July, 29; minor changes in color designation; 
		added a bunch of tiling examples in the new section "A systematic approach"
Tested:		with Mathematica 7	

In order to see the graphics this notebook must be evaluated first. All these graphics need a lot of computer resources. So perhaps it may be advisable to evaluate only one section at time, But always allow the initialization cells to be evaluated.